@@ -6,31 +6,24 @@
     <packaging>jar</packaging>
 
     <name>Podio API Client</name>
-    <version>0.8.0-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
     <description>Fork of the official Java wrapper for the Podio API</description>
     <url>http://podio.com</url>
 
     <scm>
-        <connection>scm:git:git@github.com:daniel-sc/podio-java.git</connection>
-        <developerConnection>scm:git:git@github.com:daniel-sc/podio-java.git</developerConnection>
-        <url>git@github.com:daniel-sc/podio-java.git</url>
+        <connection>scm:git:git@github.com:podio-community/podio-java.git</connection>
+        <developerConnection>scm:git:git@github.com:podio-community/podio-java.git</developerConnection>
+        <url>git@github.com:podio-community/podio-java.git</url>
       <tag>HEAD</tag>
   </scm>
-    <distributionManagement>
-        <repository>
-            <id>github</id>
-            <name>GitHub daniel-sc Apache Maven Packages</name>
-            <url>https://maven.pkg.github.com/daniel-sc/podio-java</url>
-        </repository>
-    </distributionManagement>
     <build>
         <plugins>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>2.3.2</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>11</source>
+                    <target>11</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -81,20 +74,19 @@
             <version>4.12</version>
         </dependency>
         <dependency>
-            <groupId>com.sun.jersey</groupId>
+            <groupId>org.glassfish.jersey.core</groupId>
             <artifactId>jersey-client</artifactId>
-            <version>1.19.4</version>
+            <version>2.30.1</version>
         </dependency>
         <dependency>
-            <groupId>com.sun.jersey.contribs</groupId>
-            <artifactId>jersey-multipart</artifactId>
-            <version>1.19.4</version>
+            <groupId>org.glassfish.jersey.media</groupId>
+            <artifactId>jersey-media-json-jackson1</artifactId>
+            <version>2.30.1</version>
         </dependency>
         <dependency>
-            <groupId>com.sun.jersey</groupId>
-            <artifactId>jersey-json</artifactId>
-            <version>1.19.4</version>
-            <scope>compile</scope>
+            <groupId>org.glassfish.jersey.media</groupId>
+            <artifactId>jersey-media-multipart</artifactId>
+            <version>2.30.1</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
