@@ -1,23 +1,19 @@
 package com.podio;
 
-import com.sun.jersey.api.client.ClientHandlerException;
-import com.sun.jersey.api.client.ClientRequest;
-import com.sun.jersey.api.client.ClientResponse;
-import com.sun.jersey.api.client.filter.ClientFilter;
+
+import javax.ws.rs.client.ClientRequestContext;
+import javax.ws.rs.client.ClientRequestFilter;
+import java.io.IOException;
 
 /**
  * Used to put the API in testmode
  */
-public class DryRunFilter extends ClientFilter {
+public class DryRunFilter implements ClientRequestFilter {
 
 	@Override
-	public ClientResponse handle(ClientRequest cr)
-			throws ClientHandlerException {
-		if (!cr.getURI().getPath().startsWith("/oauth")) {
+	public void filter(ClientRequestContext cr) throws IOException {
+		if (!cr.getUri().getPath().startsWith("/oauth")) {
 			cr.getHeaders().putSingle("X-Podio-Dry-Run", "steYut44");
 		}
-
-		return getNext().handle(cr);
 	}
-
 }
