@@ -1,21 +1,21 @@
 package com.podio.oauth;
 
-import javax.ws.rs.core.MultivaluedMap;
+import java.util.Map;
 
 public class OAuthRefreshTokenCredentials implements OAuthUserCredentials {
 
-	private final String refreshToken;
+    private final String refreshToken;
 
-	public OAuthRefreshTokenCredentials(String refreshToken) {
-		super();
-		this.refreshToken = refreshToken;
-	}
+    public OAuthRefreshTokenCredentials(String refreshToken) {
+        super();
+        this.refreshToken = refreshToken;
+    }
 
-	public String getType() {
-		return "refresh_token";
-	}
+    public String getType() {
+        return "refresh_token";
+    }
 
-	public void addParameters(MultivaluedMap<String, String> map) {
-		map.add("refresh_token", refreshToken);
-	}
+    public void addParameters(Map<String, String> map) {
+        map.put("refresh_token", refreshToken);
+    }
 }
