@@ -1,18 +1,18 @@
 package com.podio.user;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import javax.ws.rs.core.MediaType;
-
 import com.podio.BaseAPI;
 import com.podio.ResourceFactory;
 import com.podio.contact.Profile;
 import com.podio.contact.ProfileField;
 import com.podio.contact.ProfileFieldValues;
 import com.podio.contact.ProfileUpdate;
-import com.sun.jersey.api.client.GenericType;
+
+import javax.ws.rs.client.Entity;
+import javax.ws.rs.core.GenericType;
+import javax.ws.rs.core.MediaType;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * This area holds all the users which just includes basic operations.
@@ -30,13 +30,13 @@ public class UserAPI extends BaseAPI {
 	 */
 	public void updateUser(UserUpdate update) {
 		getResourceFactory().getApiResource("/user/")
-				.entity(update, MediaType.APPLICATION_JSON_TYPE).put();
+                .put(Entity.entity(update, MediaType.APPLICATION_JSON_TYPE));
 	}
 
 	/**
 	 * Returns the current status for the user. This includes the user data,
 	 * profile data and notification data.
-	 * 
+	 *
 	 * @return The status of the user
 	 */
 	public UserStatus getStatus() {
@@ -46,7 +46,7 @@ public class UserAPI extends BaseAPI {
 
 	/**
 	 * Returns the profile of the active user
-	 * 
+	 *
 	 * @return The profile for the user
 	 */
 	public Profile getProfile() {
@@ -56,16 +56,15 @@ public class UserAPI extends BaseAPI {
 
 	/**
 	 * Returns the field of the profile for the given key from the active user.
-	 * 
+	 *
 	 * @param field
 	 *            The field to return the values for
 	 * @return The values for the given field
 	 */
 	public <T, R> List<T> getProfileField(ProfileField<T, R> field) {
 		List<R> values = getResourceFactory().getApiResource(
-				"/user/profile/" + field.getName()).get(
-				new GenericType<List<R>>() {
-				});
+				"/user/profile/" + field.getName()).get().readEntity(new GenericType<List<R>>() {
+		});
 
 		List<T> formatted = new ArrayList<T>();
 		for (R value : values) {
@@ -78,18 +77,18 @@ public class UserAPI extends BaseAPI {
 	/**
 	 * Updates the fields of an existing profile. All fields must be filled out,
 	 * as any fields not included will not be part of the new revision.
-	 * 
+	 *
 	 * @param profile
 	 *            The updated profile
 	 */
 	public void updateProfile(ProfileUpdate profile) {
 		getResourceFactory().getApiResource("/user/profile/")
-				.entity(profile, MediaType.APPLICATION_JSON_TYPE).put();
+				.put(Entity.entity(profile, MediaType.APPLICATION_JSON_TYPE));
 	}
 
 	/**
 	 * Updates a single field on the profile of the user
-	 * 
+	 *
 	 * @param field
 	 *            The field that should be updated
 	 * @param value
@@ -99,19 +98,19 @@ public class UserAPI extends BaseAPI {
 		if (field.isSingle()) {
 			getResourceFactory()
 					.getApiResource("/user/profile/" + field.getName())
-					.entity(new ProfileFieldSingleValue<F>(value),
-							MediaType.APPLICATION_JSON_TYPE).put();
+					.put(Entity.entity(new ProfileFieldSingleValue<F>(value),
+							MediaType.APPLICATION_JSON_TYPE));
 		} else {
 			getResourceFactory()
 					.getApiResource("/user/profile/" + field.getName())
-					.entity(new ProfileFieldMultiValue<F>(value),
-							MediaType.APPLICATION_JSON_TYPE).put();
+                    .put(Entity.entity(new ProfileFieldMultiValue<F>(value),
+							MediaType.APPLICATION_JSON_TYPE));
 		}
 	}
 
 	/**
 	 * Updates a single field on the profile of the user
-	 * 
+	 *
 	 * @param field
 	 *            The field that should be updated
 	 * @param values
@@ -123,7 +122,7 @@ public class UserAPI extends BaseAPI {
 
 	/**
 	 * Updates a single field on the profile of the user
-	 * 
+	 *
 	 * @param field
 	 *            The field that should be updated
 	 * @param values
@@ -136,26 +135,26 @@ public class UserAPI extends BaseAPI {
 		} else {
 			getResourceFactory()
 					.getApiResource("/user/profile/" + field.getName())
-					.entity(new ProfileFieldMultiValue<F>(values),
-							MediaType.APPLICATION_JSON_TYPE).put();
+                    .put(Entity.entity(new ProfileFieldMultiValue<F>(values),
+							MediaType.APPLICATION_JSON_TYPE));
 		}
 	}
 
 	/**
 	 * Updates the fields of an existing profile. Will only update the fields in
 	 * the values.
-	 * 
+	 *
 	 * @param values
 	 *            The updated values for the profile
 	 */
 	public void updateProfile(ProfileFieldValues values) {
 		getResourceFactory().getApiResource("/user/profile/")
-				.entity(values, MediaType.APPLICATION_JSON_TYPE).put();
+                .put(Entity.entity(values, MediaType.APPLICATION_JSON_TYPE));
 	}
 
 	/**
 	 * Gets the active user
-	 * 
+	 *
 	 * @return The active user
 	 */
 	public User getUser() {
@@ -165,7 +164,7 @@ public class UserAPI extends BaseAPI {
 	/**
 	 * Returns the value of the property for the active user with the given
 	 * name. The property is specific to the auth client used.
-	 * 
+	 *
 	 * @param key
 	 *            The key of the property
 	 */
@@ -177,7 +176,7 @@ public class UserAPI extends BaseAPI {
 	/**
 	 * Sets the value of the property for the active user with the given name.
 	 * The property is specific to the auth client used.
-	 * 
+	 *
 	 * @param key
 	 *            The key of the property
 	 * @param value
@@ -186,14 +185,14 @@ public class UserAPI extends BaseAPI {
 	public void setProperty(String key, boolean value) {
 		getResourceFactory()
 				.getApiResource("/user/property/" + key)
-				.entity(new PropertyValue(value),
-						MediaType.APPLICATION_JSON_TYPE).put();
+                .put(Entity.entity(new PropertyValue(value),
+						MediaType.APPLICATION_JSON_TYPE));
 	}
 
 	/**
 	 * Deletes the property for the active user with the given name. The
 	 * property is specific to the auth client used.
-	 * 
+	 *
 	 * @param key
 	 *            The key of the property that should be deleted
 	 */
